Working on QueryBuilder.
